Fix: 后端 CI 在 Linux/Py3.11 下两处失败
- known_folder.GUID: Data1 由 c_ulong 改为 c_uint32（Linux x86-64 上 unsigned long 为 8 字节，导致结构体 24 而非 16 字节；Windows DWORD 恒为 32 位）。现用固定宽度类型，所有平台均为 16 字节，Windows marshalling 行为不变。
- test_binary_autosave_header: canParseLocally 断言改为依赖 Rust sidecar 可用性（CI 的 server 作业不构建 sidecar），缺失时走 missingComponent 路径，与 CI 策略一致。

+++ apps/server/app/services/known_folder.py
@@ -31,10 +31,10 @@ class GUID(ctypes.Structure):
     """
 
     _fields_ = [
-        ("Data1", ctypes.c_ulong),
-        ("Data2", ctypes.c_ushort),
-        ("Data3", ctypes.c_ushort),
-        ("Data4", ctypes.c_ubyte * 8),
+        ("Data1", ctypes.c_uint32),  # Windows GUID DWORD 恒为 32 位（非 c_ulong：Linux x86-64 上 unsigned long 为 8 字节）
+        ("Data2", ctypes.c_uint16),  # WORD
+        ("Data3", ctypes.c_uint16),  # WORD
+        ("Data4", ctypes.c_uint8 * 8),  # BYTE[8]
     ]
 
     @classmethod
@@ -44,7 +44,7 @@ class GUID(ctypes.Structure):
         if len(parts) != 5:
             raise ValueError(f"非法 GUID 字符串：{s!r}")
         data4_hex = parts[3] + parts[4]
-        data4 = (ctypes.c_ubyte * 8)(*bytes.fromhex(data4_hex))
+        data4 = (ctypes.c_uint8 * 8)(*bytes.fromhex(data4_hex))
         return cls(
             Data1=int(parts[0], 16),
             Data2=int(parts[1], 16),
